requirements.txt
Flask == 2.3.2
python-dotenv == 1.0.0
Flask-SQLAlchemy == 3.0.5
# flask-marshmallow == 0.15.0
# marshmallow-SQLAlchemy == 0.29.0
# SQLAlchemy-filters == 0.13.0
SQLAlchemy==2.0.20
marshmallow==3.20.1
# Flask-Migrate==4.0.4
# flask_cors==4.0.0
#psycopg2==2.9.7
psycopg2-binary==2.9.9
# werkzeug==2.3.7